backend/requirements.txt
Flask-CORS==3.0.10
Flask==2.0.2
Werkzeug==2.0.2
psycopg2-binary==2.9.2
python-dotenv==0.19.2